LayoutTests/platform/mac/accessibility/table-with-row-col-of-headers-expected.txt

header 1	header 2
cell 1
header 1	header 2
cell 1	cell 1
header 1	header 2
cell 1
header 1	header 2
cell 1
This tests that tables that have THs in the first row or first column will be exposed as AXTables.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.role is 'AXRole: AXTable'
PASS accessibilityController.focusedElement.role is 'AXRole: AXTable'
PASS accessibilityController.focusedElement.role is 'AXRole: AXWebArea'
PASS accessibilityController.focusedElement.role is 'AXRole: AXWebArea'
PASS successfullyParsed is true

TEST COMPLETE
